Map support (#15)
* Added map objects

AB#37

* Added simple map

* Environment parts implements StaticMapObject

* Added Map class with MapLoader

AB#22 AB#23

* unit is now integer

* PlayersController implements DynamicMapObject

* Game screen renders a given map

AB#39

* Changed wall color to cyan

* Provided support for InputController

+++ core/src/com/geometric/wars/Values.java
@@ -1,5 +1,5 @@
 package com.geometric.wars;
 
 public class Values {
-    public static final float unit = 1f;
+    public static final int unit = 1;
 }
